pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibeship-optimizer"
version = "0.1.0"
description = "vibeship-optimizer: safe, rollback-friendly optimization workflow with snapshots + compare + verification."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "OpenClaw community" }]
keywords = ["performance", "optimization", "validation", "openclaw", "ops"]

dependencies = ["PyYAML>=6.0"]

[project.scripts]
vibeship-optimizer = "vibeship_optimizer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
